Move builtins-related code to its own module
