+++ dhall/src/semantics/mod.rs
@@ -1,7 +1,9 @@
+pub mod builtins;
 pub mod core;
 pub mod nze;
 pub mod phase;
 pub mod tck;
+pub(crate) use self::builtins::*;
 pub(crate) use self::core::*;
 pub(crate) use self::nze::*;
 pub(crate) use self::tck::*;
